Refactor triangulation into a separate class
